Implement values, limits validation in Action and checks.
This adds a dependency to funcsigs which is a backport of inspect.Signature to python lower than 3.3. It is a pure python package .

Checks does not rely anymore on the presence of {} to identify driver related attribute those can be simply accessed through self (which means that one can access to parent attribute which was not possible before). The drawback is that error message is less informative I plan to fix (see comment in code).